Move routes to ./routes
