+++ src/server/routes/about/controller.test.js
@@ -1,4 +1,4 @@
-import { createServer } from '../server.js'
+import { createServer } from '#/server/server.js'
 import { statusCodes } from '#/server/common/constants/status-codes.js'
 
 describe('#aboutController', () => {
