+++ src/server/routes/health/controller.js
@@ -1,4 +1,4 @@
-import { statusCodes } from '../common/constants/status-codes.js'
+import { statusCodes } from '#/server/common/constants/status-codes.js'
 
 /**
  * A generic health-check endpoint. Used by the platform to check if the service is up and handling requests.
